--- a/ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+WKNavigationDelegate.swift
+++ b/ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+WKNavigationDelegate.swift
@@ -436,11 +436,11 @@ extension BrowserViewController: WKNavigationDelegate {
         {
           let domain = Domain.getOrCreate(forUrl: requestURL, persistent: !isPrivateBrowsing)
 
-          let shouldBlock = await AdBlockStats.shared.shouldBlock(
+          let shouldBlock = await AdBlockGroupsManager.shared.shouldBlock(
             requestURL: requestURL,
             sourceURL: requestURL,
             resourceType: .document,
-            isAggressiveMode: domain.blockAdsAndTrackingLevel.isAggressive
+            domain: domain
           )
 
           if shouldBlock, let url = requestURL.encodeEmbeddedInternalURL(for: .blocked) {