--- a/ios/brave-ios/Sources/Brave/Frontend/Browser/PageData.swift
+++ b/ios/brave-ios/Sources/Brave/Frontend/Browser/PageData.swift
@@ -22,11 +22,11 @@ struct PageData {
   /// These are loaded dyncamically as the user scrolls through the page
   private(set) var allSubframeURLs: Set<URL> = []
   /// The stats class to get the engine data from
-  private var adBlockStats: AdBlockStats
+  private var groupsManager: AdBlockGroupsManager
 
-  init(mainFrameURL: URL, adBlockStats: AdBlockStats = AdBlockStats.shared) {
+  init(mainFrameURL: URL, groupsManager: AdBlockGroupsManager = AdBlockGroupsManager.shared) {
     self.mainFrameURL = mainFrameURL
-    self.adBlockStats = adBlockStats
+    self.groupsManager = groupsManager
   }
 
   /// This method builds all the user scripts that should be included for this page
@@ -120,7 +120,7 @@ struct PageData {
     domain: Domain,
     isDeAmpEnabled: Bool
   ) async -> Set<UserScriptType> {
-    return await adBlockStats.makeEngineScriptTypes(
+    return await groupsManager.makeEngineScriptTypes(
       frameURL: mainFrameURL,
       isMainFrame: true,
       isDeAmpEnabled: isDeAmpEnabled,
@@ -130,7 +130,7 @@ struct PageData {
 
   func makeAllEngineScripts(for domain: Domain, isDeAmpEnabled: Bool) async -> Set<UserScriptType> {
     // Add engine scripts for the main frame
-    async let engineScripts = adBlockStats.makeEngineScriptTypes(
+    async let engineScripts = groupsManager.makeEngineScriptTypes(
       frameURL: mainFrameURL,
       isMainFrame: true,
       isDeAmpEnabled: isDeAmpEnabled,
@@ -139,7 +139,7 @@ struct PageData {
 
     // Add engine scripts for all of the known sub-frames
     async let additionalScriptTypes = allSubframeURLs.asyncConcurrentCompactMap({ frameURL in
-      return await self.adBlockStats.makeEngineScriptTypes(
+      return await self.groupsManager.makeEngineScriptTypes(
         frameURL: frameURL,
         isMainFrame: false,
         isDeAmpEnabled: isDeAmpEnabled,