--- a/ios/brave-ios/Sources/Brave/Frontend/Settings/Features/ShieldsPrivacy/AdvancedShieldSettings.swift
+++ b/ios/brave-ios/Sources/Brave/Frontend/Settings/Features/ShieldsPrivacy/AdvancedShieldSettings.swift
@@ -28,17 +28,21 @@ import os
   @Published var cookieConsentBlocking: Bool {
     didSet {
       FilterListStorage.shared.ensureFilterList(
-        for: FilterList.cookieConsentNoticesComponentID,
+        for: AdblockFilterListCatalogEntry.cookieConsentNoticesComponentID,
         isEnabled: cookieConsentBlocking
       )
+
+      AdBlockGroupsManager.shared.compileEnginesIfNeeded()
     }
   }
   @Published var blockMobileAnnoyances: Bool {
     didSet {
       FilterListStorage.shared.ensureFilterList(
-        for: FilterList.mobileAnnoyancesComponentID,
+        for: AdblockFilterListCatalogEntry.mobileAnnoyancesComponentID,
         isEnabled: blockMobileAnnoyances
       )
+
+      AdBlockGroupsManager.shared.compileEnginesIfNeeded()
     }
   }
   @Published var isP3AEnabled: Bool {
@@ -101,11 +105,11 @@ import os
     self.isDebounceEnabled = debounceService?.isEnabled ?? false
 
     cookieConsentBlocking = FilterListStorage.shared.isEnabled(
-      for: FilterList.cookieConsentNoticesComponentID
+      for: AdblockFilterListCatalogEntry.cookieConsentNoticesComponentID
     )
 
     blockMobileAnnoyances = FilterListStorage.shared.isEnabled(
-      for: FilterList.mobileAnnoyancesComponentID
+      for: AdblockFilterListCatalogEntry.mobileAnnoyancesComponentID
     )
 
     var clearableSettings = [
@@ -232,11 +236,11 @@ import os
       .sink { filterLists in
         for filterList in filterLists {
           switch filterList.entry.componentId {
-          case FilterList.cookieConsentNoticesComponentID:
+          case AdblockFilterListCatalogEntry.cookieConsentNoticesComponentID:
             if filterList.isEnabled != self.cookieConsentBlocking {
               self.cookieConsentBlocking = filterList.isEnabled
             }
-          case FilterList.mobileAnnoyancesComponentID:
+          case AdblockFilterListCatalogEntry.mobileAnnoyancesComponentID:
             if filterList.isEnabled != self.blockMobileAnnoyances {
               self.blockMobileAnnoyances = filterList.isEnabled
             }