--- a/ios/brave-ios/Sources/Brave/Frontend/Settings/Features/ShieldsPrivacy/FilterLists/FilterListsView.swift
+++ b/ios/brave-ios/Sources/Brave/Frontend/Settings/Features/ShieldsPrivacy/FilterLists/FilterListsView.swift
@@ -64,9 +64,7 @@ struct FilterListsView: View {
       )
     }
     .onDisappear {
-      Task.detached {
-        await AdBlockGroupsManager.shared.ensureEnabledEngines()
-      }
+      AdBlockGroupsManager.shared.compileEnginesIfNeeded()
     }
   }