--- a/ios/brave-ios/Sources/Brave/Frontend/UserContent/UserScripts/Scripts_Dynamic/ScriptHandlers/Paged/RequestBlockingContentScriptHandler.swift
+++ b/ios/brave-ios/Sources/Brave/Frontend/UserContent/UserScripts/Scripts_Dynamic/ScriptHandlers/Paged/RequestBlockingContentScriptHandler.swift
@@ -79,11 +79,11 @@ class RequestBlockingContentScriptHandler: TabContentScript {
       Task { @MainActor in
         let domain = Domain.getOrCreate(forUrl: currentTabURL, persistent: !isPrivateBrowsing)
         guard let domainURLString = domain.url else { return }
-        let shouldBlock = await AdBlockStats.shared.shouldBlock(
+        let shouldBlock = await AdBlockGroupsManager.shared.shouldBlock(
           requestURL: requestURL,
           sourceURL: sourceURL,
           resourceType: dto.data.resourceType,
-          isAggressiveMode: domain.blockAdsAndTrackingLevel.isAggressive
+          domain: domain
         )
 
         // Ensure we check that the stats we're tracking is still for the same page