<!-- Add brave-browser issue below that this PR will resolve -->
Resolves 
https://github.com/brave/brave-browser/issues/36035
https://github.com/brave/brave-browser/issues/36539

Two issues are bundled here
1. Unify engines into two (standard/general and aggressive/additional) 
2. Load from dat files during launch. Once unified the engines have to change how we load from cache. It is natural to do this using dat files instead of text files with very little additional work. Loading cached engines from text files would result in unneeded work and unneeded testing.

<!-- CI-related labels that can be applied to this PR:
* CI/run-audit-deps (1) - check for known npm/cargo vulnerabilities (audit_deps)
* CI/run-network-audit (1) - run network-audit
* CI/run-upstream-tests - run Chromium unit and browser tests on Linux and Windows (otherwise only on Linux)
* CI/run-linux-arm64, CI/run-macos-arm64, CI/run-windows-arm64, CI/run-windows-x86 - run builds that would otherwise be skipped
* CI/skip - do not run CI builds (except noplatform)
* CI/skip-linux-x64, CI/skip-android, CI/skip-macos-x64, CI/skip-ios, CI/skip-windows-x64 - skip CI builds for specific platforms
* CI/skip-upstream-tests - do not run Chromium unit, or browser tests (otherwise only on Linux)
* CI/skip-all-linters - do not run presubmit and lint checks
* CI/storybook-url (1) - deploy storybook and provide a unique URL for each build

(1) applied automatically when some files are changed (see: https://github.com/brave/brave-core/blob/master/.github/labeler.yml)
-->

## Submitter Checklist:

- [ ] I confirm that no [security/privacy review is needed](https://github.com/brave/brave-browser/wiki/Security-reviews) and no other type of reviews are needed, or that I have [requested](https://github.com/brave/reviews/issues/new/choose) them
- [ ] There is a [ticket](https://github.com/brave/brave-browser/issues) for my issue
- [ ] Used Github [auto-closing keywords](https://docs.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue) in the PR description above
- [ ] Wrote a good [PR/commit description](https://google.github.io/eng-practices/review/developer/cl-descriptions.html)
- [ ] Squashed any review feedback or "fixup" commits before merge, so that history is a record of what happened in the repo, not your PR
- [ ] Added appropriate labels (`QA/Yes` or `QA/No`; `release-notes/include` or `release-notes/exclude`; `OS/...`) to the associated issue
- [ ] Checked the PR locally:
  * `npm run test -- brave_browser_tests`, `npm run test -- brave_unit_tests` [wiki](https://github.com/brave/brave-browser/wiki/Tests)
  * `npm run presubmit` [wiki](https://github.com/brave/brave-browser/wiki/Presubmit-checks), `npm run gn_check`, `npm run tslint`
- [ ] Ran `git rebase master` (if needed)

## Reviewer Checklist:

- [ ] A security review [is not needed](https://github.com/brave/brave-browser/wiki/Security-reviews), or a link to one is included in the PR description
- [ ] New files have MPL-2.0 license header
- [ ] Adequate test coverage exists to prevent regressions
- [ ] Major classes, functions and non-trivial code blocks are well-commented
- [ ] Changes in component dependencies are properly reflected in `gn`
- [ ] Code follows the [style guide](https://chromium.googlesource.com/chromium/src/+/HEAD/styleguide/c++/c++.md)
- [ ] Test plan is specified in PR before merging

## After-merge Checklist:

- [ ] The associated issue milestone is set to the smallest version that the
  changes has landed on
- [ ] All relevant documentation has been updated, for instance:
  - [ ] https://github.com/brave/brave-browser/wiki/Deviations-from-Chromium-(features-we-disable-or-remove)
  - [ ] https://github.com/brave/brave-browser/wiki/Proxy-redirected-URLs
  - [ ] https://github.com/brave/brave-browser/wiki/Fingerprinting-Protections
  - [ ] https://github.com/brave/brave-browser/wiki/Brave%E2%80%99s-Use-of-Referral-Codes
  - [ ] https://github.com/brave/brave-browser/wiki/Web-Compatibility-Exceptions-in-Brave
  - [ ] https://github.com/brave/brave-browser/wiki/QA-Guide
  - [ ] https://github.com/brave/brave-browser/wiki/P3A

## Test Plan:
1. Test that youtube adblocking works
2. Test that youtube adblocking works during relaunch and heading directly to youtube.
3. Test https://dev-pages.brave.software/filtering/cosmetic-filtering.html
4. Test https://dev-pages.brave.software/filtering/scriptlets.html
5. Test https://dev-pages.brave.software/filtering/additional-lists.html
6. Test https://dev-pages.bravesoftware.com/filtering/text-ads.html
7. Test https://dev-pages.brave.software/filtering/scriptlets.html

## Performance:
### iPhone 8
|  | Previous | This |
|---|---|---|
| Launch | ~11s | ~3.3s (default) ~8s (all filter lists) |
| Post-launch | ~2.2s (default) ~12s (all filter lists) | ~0.2s |
| Screenshot | <img width="1522" alt="Screenshot 2024-03-27 at 11 21 16 AM" src="https://github.com/brave/brave-core/assets/909331/ecaaae39-7bd0-45ce-9669-b21413292079"> | <img width="1367" alt="Screenshot 2024-03-28 at 3 19 10 PM" src="https://github.com/brave/brave-core/assets/909331/48cb63a1-cdda-469f-8042-962bba4af61e"> |

### iPhone 15 pro
|  | Previous | This |
|---|---|---|
| Launch | ~2s | ~0.9s (default) ~2s (all filter lists) |
| Post-launch | ~0.9s (default) ~4s (all filter lists) | ~0.02s |